environments directory has been renamed to setup
