+++ test/assertions/escrowAssertion.ts
@@ -5,8 +5,8 @@ import { EscrowType } from "../../src/shared/types/escrowType"
 import { EscrowEvm } from "../../src/core/escrow/repo/evm"
 import { Fund } from "../../src/core/escrow/types"
 import { handleEvmError } from "../shared/error"
-import { Accounts } from "../shared/accounts"
-import { IAccounts } from "../interfaces/environments/IAccounts"
+import { Accounts } from "../testkits/setup/accounts"
+import { IAccounts } from "../interfaces/setup/IAccounts"
 
 
 export class EscrowAssertion {
